move logic into different files
